requirements.txt
flask
chardet
nltk
google-generativeai
python-dotenv